Add files via upload
The login page and the behind-process of it are in one file (login.php). The registration page and the behind-process of it are in one file (registration.php). I also changed the order page of the customer. Instead of directly ordering the product, the customer must add it first to his/her cart and click checkout before he/she can order the product. The missing webpage and process are also included in the following files.